Move apiCommunication
